requirements.txt
discord.py
spotipy~=2.23.0
python-dotenv~=0.20.0